auth done

+++ app/actions/deleteTodo.ts
@@ -1,13 +1,20 @@
 "use server";
 
+import { auth } from "@/auth";
 import { prisma } from "@/lib/prisma";
 import { revalidatePath } from "next/cache";
 
 export const deleteTodo = async (id: string) => {
+    const user = await auth();
+    if(!user) {
+        return { error: "Unauthorized" };
+    }
+
     try {
         const todo = await prisma.todo.delete({
             where: {
-                id: id
+                id: id,
+                userId: user?.id,
             }
         });
